@@ -1,10 +1,10 @@
 name: event_schedule_calendar
 description: develop calendar.
-version: 1.0.2
+version: 1.1.0
 repository: https://github.com/hori-imuuzak/flutter_event_schedule_calendar
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -14,7 +14,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test:
+  test: ^1.16.5
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
