@@ -2,8 +2,8 @@ import 'package:event_schedule_calendar/src/entity/palette_config.dart';
 import 'package:event_schedule_calendar/src/entity/weekday_label_config.dart';
 
 class CalendarConfig {
-  WeekdayLabelConfig weekdayLabelConfig;
-  PaletteConfig paletteConfig;
+  WeekdayLabelConfig? weekdayLabelConfig;
+  PaletteConfig? paletteConfig;
 
   final int beginWeekday;
   final String remainEventCountPattern;
