@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 class Event {
-  final DateTime dateTime;
-  final String name;
-  final Color color;
+  final DateTime? dateTime;
+  final String? name;
+  final Color? color;
 
   Event({
     this.dateTime,
