@@ -2,7 +2,7 @@ import 'package:event_schedule_calendar/src/entity/event.dart';
 import 'package:flutter/material.dart';
 
 class EventBand extends StatelessWidget {
-  final Event event;
+  final Event? event;
 
   EventBand({
     this.event,
@@ -15,11 +15,11 @@ class EventBand extends StatelessWidget {
       padding: const EdgeInsets.symmetric(vertical: 2, horizontal: 4),
       margin: const EdgeInsets.only(bottom: 4, right: 4, left: 0.5),
       decoration: BoxDecoration(
-        color: event.color ?? Colors.green,
+        color: event!.color ?? Colors.green,
         borderRadius: BorderRadius.circular(2),
       ),
       child: Text(
-        event.name,
+        event!.name!,
         maxLines: 1,
         overflow: TextOverflow.ellipsis,
         style: TextStyle(
