@@ -2,4 +2,4 @@ export 'src/entity/calendar_config.dart';
 export 'src/entity/event.dart';
 export 'src/entity/palette_config.dart';
 export 'src/entity/weekday_label_config.dart';
-export 'src/view/calendar_view/calendar_view.dart';
+export 'src/view/calendar_view/event_schedule_calendar.dart';
