@@ -3,6 +3,6 @@ import 'package:event_schedule_calendar/src/ext/same_date.dart';
 
 class EventService {
   List<Event> filter(List<Event> eventList, DateTime dateTime) {
-    return eventList.where((e) => dateTime.isSameDay(e.dateTime)).toList();
+    return eventList.where((e) => dateTime.isSameDay(e.dateTime!)).toList();
   }
 }
